+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/EmployeeTablePanel.java
@@ -22,13 +22,12 @@ import is.codion.framework.demos.chinook.model.EmployeeTableModel;
 import is.codion.swing.framework.model.SwingEntityTableModel;
 import is.codion.swing.framework.ui.EntityTablePanel;
 
+import javax.swing.JComponent;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JSplitPane;
-import java.awt.BorderLayout;
 
 import static is.codion.swing.common.ui.component.Components.splitPane;
-import static is.codion.swing.common.ui.layout.Layouts.borderLayout;
 import static is.codion.swing.framework.ui.EntityTree.entityTree;
 
 public final class EmployeeTablePanel extends EntityTablePanel {
@@ -38,18 +37,14 @@ public final class EmployeeTablePanel extends EntityTablePanel {
   }
 
   @Override
-  protected void layoutPanel(JPanel tablePanel, JPanel southPanel) {
-    setLayout(borderLayout());
-    EmployeeTableModel employeeTableModel = tableModel();
-    JSplitPane splitPane = splitPane()
+  protected void layoutPanel(JComponent tableComponent, JPanel southPanel) {
+    super.layoutPanel(splitPane()
             .orientation(JSplitPane.HORIZONTAL_SPLIT)
             .continuousLayout(true)
             .oneTouchExpandable(true)
             .resizeWeight(0.65)
-            .leftComponent(tablePanel)
-            .rightComponent(new JScrollPane(entityTree(employeeTableModel.treeModel())))
-            .build();
-    add(splitPane, BorderLayout.CENTER);
-    add(southPanel, BorderLayout.SOUTH);
+            .leftComponent(tableComponent)
+            .rightComponent(new JScrollPane(entityTree(((EmployeeTableModel) tableModel()).treeModel())))
+            .build(), southPanel);
   }
 }