+++ changelog.md
@@ -2,3 +2,6 @@ Codion Change Log
 ==================
 
 ## 0.17.18-SNAPSHOT
+- Text.padString() deprecated for removal along with Text.ALIGNMENT, Text.leftPad() and rightPad() added.
+- EntityTable.layoutPanel(), tablePanel parameter replaced with tableComponent, no longer restricted to JPanel.
+- Jasper Reports and Apache Commons Logging upgraded.